refactor : 외부 ID 로 장소 상세 찾기 이슈 21번 적용
